requirements.txt
streamlit
transformers
torch
torchvision
edge-tts
plotly
pandas
python-dotenv
altair
